pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "bleach>=6.2.0",
    "channels[daphne]>=4.2.2",
    "dj-database-url>=2.3.0",
    "django-ninja>=1.4.1",
    "django-ninja-extra>=0.30.0",
    "django-split-settings>=1.3.2",
    "google-genai>=1.14.0",
    "httpx>=0.28.1",
    "numpy>=2.2.5",
    "pre-commit>=4.2.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.4",
    "pydantic-ai>=0.1.11",
    "sentry-sdk>=2.27.0",
    "structlog>=25.3.0",
    "validators>=0.35.0",
]
